LayoutTests/media/media-source/media-managedmse-streaming.html

<!DOCTYPE html> <!-- webkit-test-runner [ ManagedMediaSourceEnabled=true MediaSourceEnabled=true ] -->
<html>
<head>
    <title>managedmediasource-memoryPressure</title>
    <script src="../../media/media-source/media-source-loader.js"></script>
    <script src="../../media/video-test.js"></script>
    <script src="../../media/utilities.js"></script>
    <script>
    var loader;
    var source;
    var sourceBuffer;
    var gotBufferedChange = false;
    var bufferedStart;

    function loaderPromise(loader) {
        return new Promise((resolve, reject) => {
            loader.onload = resolve;
            loader.onerror = reject;
        });
    }

    window.addEventListener('load', async event => {
        try {
            findMediaElement();

            let manifests = [ 'content/test-opus-manifest.json', 'content/test-vorbis-manifest.json', 'content/test-48khz-manifest.json', 'content/test-xhe-aac-manifest.json' ];
            for (const manifest of manifests) {
                loader = new MediaSourceLoader(manifest);
                await loaderPromise(loader);
                if (ManagedMediaSource.isTypeSupported(loader.type()))
                    break;
            }

            source = new ManagedMediaSource();
            run('video.src = URL.createObjectURL(source)');
            await Promise.all([waitFor(source, 'sourceopen'), waitFor(source, 'startstreaming') ]);
            waitFor(video, 'error').then(failTest);

            run('sourceBuffer = source.addSourceBuffer(loader.type())');

            run('sourceBuffer.appendBuffer(loader.initSegment())');
            await waitFor(sourceBuffer, 'update');

            var done = false;
            waitForEventOnceOn(source, 'endstreaming', () => { done = true; });

            do {
                sourceBuffer.appendBuffer(loader.mediaSegment(0));
                await once(sourceBuffer, 'update');
                sourceBuffer.timestampOffset = sourceBuffer.buffered.end(sourceBuffer.buffered.length - 1);
            } while (!done);

            waitForEventOnceOn(source, 'startstreaming', endTest);

            sourceBuffer.remove(video.currentTime, video.currentTime + 5);
        } catch (e) {
            failTest(`Caught exception: "${e}"`);
        }
    });
    </script>
</head>
<body>
    <video controls></video>
</body>
</html>
